feat: ContainFeedDetailCoordinator로 공통 함수 분리
