+++ burstcamp/burstcamp/Coordinator/Base/Coordinator.swift
@@ -6,12 +6,13 @@
 //
 
 import Combine
+import SafariServices.SFSafariViewController
 import UIKit
 
 protocol Coordinator: AnyObject {
     var childCoordinators: [Coordinator] { get set }
     var navigationController: UINavigationController { get set }
-    var cancelBag: Set<AnyCancellable> { get }
+    var cancelBag: Set<AnyCancellable> { get set }
 
     init(navigationController: UINavigationController)
 }
@@ -32,4 +33,44 @@ protocol NormalCoordinator: Coordinator {
 
 protocol TabBarChildCoordinator: Coordinator {
     func start(viewController: UIViewController)
+}
+
+protocol ContainFeedDetailCoordinator: Coordinator { }
+
+extension ContainFeedDetailCoordinator {
+    func moveToBlogSafari(url: URL) {
+        let safariViewController = SFSafariViewController(url: url)
+        self.navigationController.present(safariViewController, animated: true)
+    }
+
+    func sinkFeedViewController(_ feedDetailViewController: FeedDetailViewController) {
+        feedDetailViewController.coordinatorPublisher
+            .sink { [weak self] event in
+                switch event {
+                case .moveToBlogSafari(let url):
+                    self?.moveToBlogSafari(url: url)
+                }
+            }
+            .store(in: &cancelBag)
+    }
+
+    func prepareFeedDetailViewController(feed: Feed) -> FeedDetailViewController {
+        let feedDetailViewModel = FeedDetailViewModel(feed: feed)
+        let feedScrapViewModel = FeedScrapViewModel(feedUUID: feed.feedUUID)
+        let feedDetailViewController = FeedDetailViewController(
+            feedDetailViewModel: feedDetailViewModel,
+            feedScrapViewModel: feedScrapViewModel
+        )
+        return feedDetailViewController
+    }
+
+    func prepareFeedDetailViewController(feedUUID: String) -> FeedDetailViewController {
+        let feedDetailViewModel = FeedDetailViewModel(feedUUID: feedUUID)
+        let feedScrapViewModel = FeedScrapViewModel(feedUUID: feedUUID)
+        let feedDetailViewController = FeedDetailViewController(
+            feedDetailViewModel: feedDetailViewModel,
+            feedScrapViewModel: feedScrapViewModel
+        )
+        return feedDetailViewController
+    }
 }
